Use attached directory as default path.
Changed STodoTarget.assign_attachments such that if a directory, d, occurs
in the list of attachments, that directory, d, is used as the last choice
for the path of any relative-path files that occur later in the list - that
is, if the file is not found in either Configuration.instance.user_path or
the current directory, use 'd' as the path for the file if a file with that
name is found in 'd'.

Also: small bug fix in Configuration.config_file_settings:
-    lines = cfgfile.read.split("\n")
+    lines = cfgfile.read.split("\n").select do |l| l !~ /^\s*#/ end
I.e., filter out lines starting with " *#" as comments.