Changed "Reminder"s to use UTC.
The time fields (which are date/times, not just times) of the Reminder
descendants (PeriodicReminder and OneTimeReminder) were changed to
be initialized to UTC at creation time, so that they are now saved
to the "database" as UTC times. A change was also made to report (e.g.,
for report output and notifications) the times in the local time zone
(via Time.getlocal). This will - likely - fix the time-zone-change
related bug re sending reminders (the bug being that the yearly change
from standard time to "daylight savings" time can cause reminder
notifications to be an hour late, or an hour early for the reverse
scenario (e.g., CDT to CST).