Attachment processing: find/execute '.stodo-shell'
Attachment processing: If the "current" attachment is a directory and
if there is an executable file in that directory called .stodo-shell,
execute it.

Also:

README.md: Updated to match recent changes, such as the new attachments-
processing feature.

New 'version' command and supporting constant/method in Configuration.

STodoAdministrator: bug fix in 'perform_data_backup': This line:
    config = Configuration.instance
was in the wrong place.

Some small improvements to tests re reporting results.