Added -r (recursive) option to delete command.
Also:

STodoTarget.initialize_copy: Bug fix: Added guard to avoid nil-access
error regarding attachments when cloning an item: First tried this change:
-    orig.attachments.each do |o|
+    ! orig.attachments.nil? && orig.attachments.each do |o|

but a test of the 'clone' command on an item with attachments caused
an apparent infinite loop, or recursion - I had to interrupt the process.
So I changed the attachments copy logic to:
    @attachments = []
    if ! orig.attachments.nil? then
      orig.attachments.each do |o|
        @attachments << Attachment.new(o.path)
      end
    end
This worked fine with the same test. The first change was also made
for categories and references - i.e.:
    ! orig.categories.nil? && orig.categories.each do |o|
    ! orig.references.nil? && orig.references.each do |o|
It hasn't been tested with items with categories and/or references,
so the same problem might occur for these attributes.