"stodo add": initiate_new_targets; descr append
STodoManager: 'add_new_targets', which implements the "stodo add"
functionality, now includes the previously missing behavior of
calling 'initiate' on any new STodoTarget objects (which calls:
    send_initial_notifications(client)
    set_initial_calendar_entry calendar
). That is, the new email that was supposed to be sent to the user
for a new STodoTarget was not occurring with "stodo add" (It was
occurring using the "stodo temp" + "stodo new" method of creating
a new STodoTarget.) - This is now fixed.

Added an "append-to-description" option: -ad, to the 'stodo change'
command to allow appending a string to the description instead of
replacing it.

Also:
Began work on providing the option to not rely on an external
calendar program for appointments. And:
.stodo_utilities: swconf: Switch to a different config path.