Allow "disabled" mailer, doc update, alarm
mailer.rb.send_message email: If the emailtemplate setting in the
config file is void or empty, do not execute the command to send the
email. Allows user to disable use of email notifications.

README.md: Updated re some of the latest user-facing changes and
more complete/up-to-date instructions on installing and configuring
stodo.

.stodo_utilities: alarm(): "Set an alarm at the specified time."