Choice of configurable "logpath" or "syslog"
Added option to configure use of syslog for logging, as well as, for
file-based logs, the path in which to place the log files. Added
documentation for this in README.md

Also, for the administrative report "stodo settings", added a line that
reports whether assertions (DBC) are enabled or disabled.