Remove invariant for ScheduledEvent
ScheduledEvent: Removed invariant (and call of it in 'final_reminder').
The invariant is triggered when an invalid date is specified, which causes
the ScheduledEvent's date_time to be 'nil', thus violating the invariant.
However, without the invariant the ScheduledEvent is created successfully
while at the same time an error/warning is logged about the invalid date.
This behavior seems (slightly) better/more-conventinet-for-the-user.
She can check the log to see why there is no date and then add a date
by running 'stodo change' and specifying a valid date. (Note that when
ruby_contracts is enabled (ENABLE_ASSERTION=1), the invariant is triggered
via the postcondition of 'initialize' in STodoTarget instead of in
ScheduledEvent.final_reminder.)