fix: reminder bug and Mailer bug
ScheduledEvent: Guard against violating the precondition of
OneTimeReminder.new via a nil date_time - i.e.:
   def final_reminder
-    if @final_reminder == nil then
+    if @final_reminder == nil && date_time != nil then
       @final_reminder = OneTimeReminder.new(date_time)
     end

Mailer: Fix problem with odd characters, such as single/double quotes,
in subject causing external email command to fail. Fixed by passing an
array instead of a string as the first argument to IO.popen. (The subject
becomes one of the elements of the array, which does not expose the ',
", etc. characters to the UNIX shell when the mail command is invoked.)

Removed Gemfile.lock.