Refactor of reminder scanning/creation
The new class PeriodicReminderConstructor is now used to construct a
PeriodicReminder based on the period_spec, date_time, and period_count
from the results given by the ReminderSpecScanner passed to
PeriodicReminderConstructor's constructor. The ReminderSpecScanner is
now also used (despite the name) to obtain the 'date_time' (from the
spec given to ReminderSpecScanner) that is passed to the
OneTimeReminder's constructor. PeriodicDateParser is no longer used.

Removed obsolete code, moved code around, and other typical changes that
tend to occur during a refactoring.