ReminderSpecScanner: array access bug fix
ReminderSpecScanner: In 'initialize' and 'date_type_for', needed to
add a guard condition (word_groups.count > 1) to prevent to prevent
access to nonexistent elements, such as word_groups[1][0] that
occurs when there is no period type - prevent processing for
PeriodicReminders, since in this case it will be a OneTimeReminder.

Minor reminder test change.