Adaptation to use redis with Logger, etc.
redis_setup.rb: Added 'setup_redis' method to optionally, based on
arguments, configure redis as the output device for the global $log
variable (the Logger), set $logkey (which is used as the key for the
redis stream (via xadd) to be used for logging) based on the 'service_name'
argument, and set a global $redis variable to be used for debugging.

report.rb: Added code for testing the Logger ($log) with redis
configured as the output device.