admin (redis-based) logging enhancements
redis_log_config: New RedisLogConfig class to replace the redis_setup.rb
hack. It sets (global) $log to an instance of Logger that is configured
to use the RedisLoggerDevice (which it creates) so that, e.g.,
$log.warn ..., $log.error, etc., calls are logged to the redis
database/stream. The key used for logging (created in the line:
    self.log_key = "#{service_name}#{$$}"
) is stored by the line:
    admin_broker.queue_messages("#{service_name}-entries", log_key)
e.g.: a log_key of 'report299374' is stored as a redis list entry
with the key 'report-entries' so that the list of log keys can be
retrieved later in order to view log entries.

Configuration: Changes to accommodate administrative logging:
'service_name' and 'debugging' attributes added to be used for logging
service_name is used as the basis of the key for the redis stream to
which logging is appended. Since Configuration creates the
RedisLogConfig to set up the logging, 'service_name' and 'debugging'
need to be set via class variables (e.g., by calling
Configuration.service_name= to set the service_name) by top-level
objects. The following classes and files were changed to accomplish this:
  STodoAdministrator
  STodoManager
  initial-processing.rb
  ongoing-processing.rb
  stodo-mgmt.rb
  stodo-template.rb
  report.rb

Changes to ConfigTools to accommodate the new 'redis' logging type and
the upcoming redis-based-database change (which adds a new
concept of "database type"):
+  DB_TYPE_TAG          = 'database_type'
+  REDIS_TYPE_TAG       = 'redis'

new classes: ObjectInfo, RedisLogConfig