logging enhancements: log-key list, etc.; bug fix
New 'stodo report' option, lklist: a list of log categories used to group
log keys according to a "category" (report, management, etc.). The
user's username is prepended to each element to prevent clashes when
there are multiple users.  Use this by first running:
  stodo rep lklist
to get the list. Then choose one or more of the reported categories
to get a list of log keys available for that category - e.g.:
  st rep logkey user.initial-processing
Then choose one or more of the resulting log keys to output the log
entires for those keys - e.g.:
  st rep logmsg user.initial-processing.20250420.220751.222645100

Began work on a request-id concept for use in grouping certain log
entries (for example, from a small set of related operations run
sequentially) together. A request-id can be used to retrieve all
log entries for the entire set of operations.

Bug fix - reportmanager.rb (was turning arg into a hash):

   def report_logmsgs(criteria)
     config = Configuration.instance
-    puts config.log_config.log_messages(key: criteria.handles[0])
+    puts config.log_config.log_messages(criteria.handles[0])
   end

Added 'config' (Configuration) as an argument to RedisLogConfig.new
so that the RedisLogConfig can use the 'user' attribute of config
for the user value now used for log-related keys.