logging-related transactions
From the top-level view, added four logging-related "transaction"
operations - from the help menu from the main 'stodo' script:
  starttran          Start a transaction
                     Note: transactions are currently only used for
                     grouping log entries together for later retrieval.
  endtran            End the current transaction.
  tranmsgs <tr-id>   Output all messages that were logged during transaction
                     with id <tr-id> (default: the last transaction)
  tranids            Output all transaction ids in the database.

In the ruby code, high-level management logic involved changes to
  report.rb, stodo-mgmt.rb, ReportManager, STodoManager
to initiate the above operations.

A new class, TransactionLog (might need a name change), implements most
of the low- and mid-level transaction logic.  RedisLogConfig creates
and configures (via arguments to 'new') the TransactionLog.

RedisMessageBroker: Used for much of the actual logging.  Added some
preconditions (e.g., that the 'key' argument is not nil) and other
small improvements.

RedisLoggerDevice: The 'write' method was changed to use the
TransactionLog, which was added as an attribute and is passed to
the constructor, to associate the stream key used to write the
log message (redis_log.send_message) with the current transaction.