Ensure most commands run within a "transaction".
TransactionLog: Added and enforced the following postcondition to
initialize/new (in a transaction unless the environment variable
STODO_SUPPRESS_TR is set):
  post :in_transaction_unless_suppressed do
    implies(ENV[SUPPRESS_TRANSACTION].nil?, in_transaction)
  end
Enforced by writing and calling 'wrap_transaction':
    if ENV[SUPPRESS_TRANSACTION].nil? then
      wrap_transaction
    end
Also added precondition to start_transaction that we are not in a
transaction. And added appropriate postcondition to start_transaction
and end_transaction.

STodoManager: changed 'start_transaction' to conform to the
above-mentioned precondition.

stodo: Set the STODO_SUPPRESS_TR environment variable to 1 to conform to
the above change in TransactionLog for 'settings', 'start-transaction',
and 'end-transaction':
+  STODO_SUPPRESS_TR=1 run $admin settings
+    STODO_SUPPRESS_TR=1 do_edit start-transaction
+    STODO_SUPPRESS_TR=1 do_edit end-transaction