Fix: don't report current open trx as previous one
TransactionManager (previously TransactionLog): Implemented the above
fix by means of a new attrubute: @cached_previous_queued_transaction.
In 'wrap_transaction', if we're not currently in a transaction, set
it to 'previous_queued_transaction' (a new convenience function)
before start_transaction is called.

Change name of class TransactionLog to TransactionManager.

Clean up a bit of unused code.