redis database: writing, updating, etc.
Some structural changes were needed to fit the paradigm shift of using
redis - update individual keys/objects/... - instead of the old
YAML::Store-based method of saving all the data to one file.
Especially, the pattern of reading all of the objects in the database
into memory and writing them all back even if just one object was
changed should no longer be followed. To help facilitate this change
a RedisSTodoManager class was created, which inherits from STodoManager.
It redefines some methods to avoid this read-everything/write-everything
behavior. This includes avoiding use of the existing_targets attribute,
which holds all of the object read from the database. The needed work
is perhaps about 50% complete. All of the tests in src/test/runtests
succeed except for the status-change test. However, the tests don't
cover everything. To make use of the RedisSTodoManager,
a new method: new_stodo_manager is to be called to "manufacture" the
right type based on whether or not the database_type is 'redis' -
i.e., either a STodoManager is created (not redis) or a
RedisSTodoManager.

A few other noteworthy changes:

Added an 'app_name' setting, which is prepended, along with the user
name, to all keys stored in the redis database.

Removed a couple irrelevant methods from RedisDBConfig that remained
from copying from RedisLogConfig.

Small improvements to RedisMessageBroker.

RedisBasedDataManager: Added methods:
    target_for
    store_target
    []
    delete
The last two were added to allow the RedisBasedDataManager to behave
like an array in order to allow it to take the place (in some cases)
of the 'existing_targets' array that is so often used.
'store_targets' now uses 'store_target'.