Oj.dump fix, fix STodoTarget state change (redis)...
Fixed the "nested too deep" error that was occurring in
when a state change was made (i.e., changing an existing
STodoTarget and then writing it back to the (redis) database)
by adding code to the 'prepare_for_db_write' of STodoTarget
that removes or "nil"s out unneeded attributes - the @db
attribute was causing the nested too deep error.

In ReportManager, the redis switch caused (among other things) the
code used for lookup for item state changes to fail - changed method:
   targets_for_criteria
The change made will break all lookups done by 'targets_for_criteria'
except those based on the item handle. Changed create_compare_methods
to do the expected lookup based on the handle rather than the
target/item used by the previous version - i.e.:
    hnd_cmp = lambda {|h, crit|
        r = crit.handle_exprs.any? do |e|
          result = h =~ /#{e}/
          result
        end
        r
    }
The other lambdas now won't work, since they remain "target-based", not
"handle-based", which is, of course, why 'targets_for_criteria' is now
broken except for handles. I.e., Some work needs to be done here to
fit into the new handle-based paradigm.
(I may decide to not implement any changes needed to keep the old,
marshalling/file-based, version working with respect to this logic.)
Added a 'selected_targets' method to STodoManager and its descendant,
RedisSTodoManager, (a handle-based lookup that "yield"s to a block)
that is used in the new version of 'targets_for_criteria'.

Implemented changes (for the new redis-based database) to
ensure that an item that is changed gets updated in the database.
The file-based version simply read in all items/targets, changed one
or more of them, and wrote them all back into the "database" (i.e.,
the one file that is used). Some highlights of these changes (without
taking the time to explain the logic):
  in STodoTarget:
    def force_update (in STodoTarget, called by STodoTargetEditor.modify_state)
    def update
    def remove
    calls to 'update' in STodoTarget methods that change 'self'.
    a 'db' attribute with public 'db=' method
  calls to (the relatively new) 'update_database' in STodoManager
  RedisBasedDataManager:
    def set_db, which uses STodoTarget's db= method.
  etc.