precondition fix: STodoTarget.modify_fields
The precondition of 'modify_fields' for STodoTarget objects was changed:
  -  pre 'target_list valid' do |spec, target_list|
  -    ! target_list.nil? && target_list.is_a?(Hash)
  -  end
  +  pre 'target_list valid' do |spec, target_list| ! target_list.nil? end
The 'target_list.is_a?(Hash)' requirement is obsolete - target_list can
now be a RedisBasedDataManager, which, for the purposes of this
application behaves like a hash table.
This change allows the test 'add-change-delete-etc-101' to succeed in
the "change" part, which relies on 'modify_fields'.

All other changes in this commit were simply removal of lines for
debugging that match:
^.*binding.irb