references test fix, transaction status
Got references test working (with redis database) - ../test/references-101.
The test creates 3 items to reference to and a 4th item that will hold
those references (the 'references' field). It was failing because the
method called that adds the references (check_references) had the
following precondition:
  pre 'extargets is-hash' do |spec| spec.existing_targets.is_a?(Hash) end
that fails when 'spec.existing_targets' is not a hash table but -
in the redis version - a RedisBasedDataManager, which provides
hash-table methods. Fixed by changing the precondition to:
  pre 'extargets is-hash' do |spec|
    spec.existing_targets.respond_to?(:has_key?)
  end
and implementing the 'has_key' function in RedisBasedDataManager.

stodo: Added a 'transaction status' command/query. Implemented in
STodoAdministrator.