Fixed broken search lambdas in ReportManager.
ReportManager: Updated all of the "lamdba"s created in
create_compare_methods to take a handle as the first block
argument instead of a target. It was broken with the change
in the "redis" version to base most searches on "handles" (keys)
rather than "targets" ({STodoTarget}s).

RedisSTodoManager.selected_targets - bug fix:
-    result = handles.keys.map { |h| @data_manager.target_for(h) }
+    result = handles.map { |h| @data_manager.target_for(h) }