Refactor STodoTarget and RedisSTodoManager
STodoManager: Replaced two lines in 'add_new_targets' with calls
to new hook methods ('store_target' and 'target_for'). These two
lines were the only ones that differed from the 'add_new_targets'
implementation in the child class RedisSTodoManager. This made
it possible to delete the 'add_new_targets' method in
RedisSTodoManager by implementing ('store_target' and 'target_for')
with behavior specific to that class (i.e., behavior needed for
the redis implementation). This change also allowed for the
deletion of 'add_child' in RedisSTodoManager, since the only line
that differed there from the version in STodoManager could be
changed to use the new 'target_for' - i.e.:
-        candidate_parent = self.existing_targets[p]
+        candidate_parent = target_for(p)

Also removed several obsolete methods in RedisSTodoManager that
were commented out (which were originally copied from STodoManager
to work with in case a different implementation was needed).
And removed 'process_targets' method from RedisSTodoManager, which
was not commented out and was exactly the same as the version in
STodoManager.