change-parent test; settings include redis
'stodo settings' now shows the database_type and log_type, which
indicate if its configured as 'redis' in the config file. In
Configuration, changed the 'settings' method to return an array
of two-element arrays to allow the order of the settings to be
coded in that method. Added a private 'settings' attribute to
eliminate the confusion of using a local 'settings' variable used
for initialization of the Configuration's attributes/state.
Changed the name of the 'settings' method to 'settings_report' to
not conflict with the new 'settings' attribute. In
STodoAdministrator.print_settings, made the needed changes
based on the above description of changes to the settings/settings_report
method.

Added a change-parent test.