tests: fix delete_test name clash.
The name clash is that 'delete_test' is defined in two places
for some of the tests - i.e., 'delete_test' is defined in
'tools'. It is also defined in: add-change-delete-etc-101,
add-change-delete-etc-102, attachments-101. Changed the
name of the function in those 3 files to remove the clash.
(I am not sure why this problem did not show up before.)

Also, in chparent-101, started a logging transaction before
the tests and ended it afterward, for a potential extra
debugging tool on test failure. (All of the tests "pieces"
will be logged together under one transaction.) To-do:
Add this to the other tests.