change-handle test, with STodoTarget fix
The new test file, test/chhandle-101, creates a parent item and
then an item assigned as child to the parent. Then it calls
'stodo chhandle ...' to change the child's handle and check if
the parent's children were updated appropriately. On the first
test run the update did not occur. The failure is due to the switch
to the redis database. Added code to change_handle in STodoTarget
to ensure that the parent is updated in the database with respect
to the child's handle change. Note the change in the 'update'
method that ensures that the parent will "prepare" its children
before being saved (preventing a "nested too deep" error from Oj):
+      self.descendants.each do |d|
+        d.prepare_for_db_write
+      end