RedisBasedSet: type for STodoTarget's children
For the configuration that uses the redis database, RedisBasedSet
replaces Set as the type for STodoTarget's children, allowing
just the handle of each child to be stored in the database rather
than the entire STodoTarget object. When finished, this will fix
problems caused by some STodoTargets being stored more than once
in the database - i.e., as an individual object and as one of its
parents children.

The configuration setup was changed to configure a "factory" method
(lambda) to use to "manufacture" the right type for new STodoTarget's
children - Set for file-based database and RedisBasedSet for the
redis database. In Configuration, the method
stodo_target_child_container_factory provides this factory and is
used by the TargetBuilder in its constructor. This requires a new
parameter in TargetBuilder's constructor, the config (Configuration)
object.

In order to behave (somewhat) like a Set, RedisBasedSet includes
Enumerable and implements the 'each' method (See:
https://avdi.codes/why-you-shouldnt-inherit-from-rubys-core-classes-and-what-to-do-instead/
), and also implements several methods found in Set, such as
'<<', 'empty?', and 'delete'.

This is not finished yet. The tests reveal a few problems still and
there may be other issues that don't show up in the tests.