redis: Fix for clear_descendants
STodoTarget: In 'remove_descendants', on the line:
      @children = new_childlist
'new_childlist' is an Array, so @children ended up being an array,
which for the redis version is the wrong type. The type should be
RedisBasedSet. Fixed by changing the line to:
      @children.merge(new_childlist)
and implementing 'merge' in RedisBasedSet. Also, needed this guard
in 'set_db' to prevent init_handles from always being called:
    if ! defined? handles || handles.nil? then
If 'set_db' is called when 'handles' is not empty, the previous
behavior ended up causing 'handles' to be cleared, causing a bug
in which a STodoTarget is incorrectly saved to the children with
no children.