file-to-redis migration conversion
In STodoAdministrator.migrate: git diff shows what was added/changed:
     targets = old_data_manager.restored_targets
-    config.data_manager.store_targets(targets)
+    data_manager = config.data_manager
+    converted_targets = data_manager.converted_from_legacy(targets)
+    data_manager.store_targets(converted_targets)
RedisBasedDataManager: 'converted_from_legacy' and 'converted' converts
a file-based target to a redis target. The key is this line in
'converted':
    target.children = RedisBasedSet.new(self)
that replaces the 'children' data structure (Array or Set) with a
RedisBasedSet. Thus all targets (STodoTarget) read from:
    targets = old_data_manager.restored_targets
are stored as "converted" targets at:
    data_manager.store_targets(converted_targets)
This addition to STodoTarget was also required:
  attr_writer   :children
to allow the RedisBasedDataManager to set 'children' to a RedisBasedSet:

For the file-based version, this extension of Set and Array was needed:
+# Adapt Array class for non-redis version (@children)
  +class Array
  +  def set_db(db)
  +    # dummy
  +  end
  +
  +  def prepare_for_db_write
  +    # dummy
  +  end
  +end
Specifically, 'prepare_for_db_write' is now called on the @children
attribute of STodoTarget, which is needed by RedisBasedSet in the
redis version. So the dummy is needed in the file-based version.

Cleanup: Removed old unused method in Configuration:
  def template_target_builder