Add assert re Reminder in STodoTarget
STodoTarget.send_ongoing_notifications: The loop over the
STodoTarget's reminders (reminders.each do |r| ...) can
sometimes encounter a "nil" reminder, which would be left
over from before a bug fix in which nil reminders will
not occur. This would cause execution to fail with an
exception. Instead of checking for nil reminders before
executing 'if r.is_due? then rems << r end', added an
assert on the line before:
      assert('r not nil') { r != nil }
since, with a valid database, this will not occur. "Fixed"
the problem that triggered the exception by removing the
offending STodoTarget (an item created for a test case).