Make 'stodo temp ...' more robust.
stodo: 'make_template': Enforce that the 'type' (1st) argument is
valid - i.e., is one of: task memo appointment event project.

In StubbedSpec.set_fields, when no handle is specified in the
"stodo temp ..." command, options.handle will be nil, which
causes a Nil object access exception later on.
Changed to check for that, as well as nil email, due_time, etc.