Changed/improved multi-line marker for editing.
Changed where the '^Gendfield^G' string (which marks the end
of a multi-line field when outputting a STodoTarget for editing)
is generated - from the bash script .stodo_utilities to the
actual ruby code. The new STODO_MULTIL_END_FIELD environment variable
tells the ruby code to output the endfield string. Changed the
appropriate line (in todo-corr) to set that variable:
    STODO_MULTIL_END_FIELD=1 stodo rep comp $handle| awk '
The new method 'multiline_end_marker_for' in SpecTools generates
the endfield line, based on the label type (description or comment).
This is called by 'to_s' in STodoTarget.