Fix remove_descendant bugs.
STodoTarget.remove_descendant: redis: In the remove-desc-101 test, one
of the grandchildren ('childE') was being removed (when 'parentB'
was the descendant to be removed) instead of being reparented
to 'ancestorA'. The fix involved changes related access of the
database (STodoTarget's 'db' attribute), specifically, in 'update',
in which 'self' is stored in the database, calling 'db=' to reset the
'db' attribute to 'tmp_db' was replaced with '@db = tmp_db' to
bypass a side effect in 'db='. And added the use of a @no_update
attribute, used by 'detach', to, when the descendant is being
"detached", suppress the database update when 'update' is called
by methods called by 'detach'. And in 'set_fields':
      self.parent_handle = spec.parent
was replaced with:
      @parent_handle = spec.parent
to avoid the 'update' call in 'parent_handle='.

In test clear-desc-101, fixed the title for the test item.