STodoTarget: database state integrity
STodoTarget: To help ensure the integrity of instances of this
class and descendants, added a 'connected_to_database' boolean
attribute and added the following rule to the class invariant:
    @connected_to_database == ! db.nil?
Also made 'prepare_for_db_write' private, since it affects the
above rule and since it was intended to be called just before
a database update so that '@db' (and other attributes) will not
get saved, and for '@db' to be restored right after the update.
The 'update' method, which is protected, does this and is the
only place where prepare_for_db_write gets called, except for
      t.send(:prepare_for_db_write)
in YamlStoreBasedDataManager (overriding the "privacy" of
prepare_for_db_write).
Also changed 'remove_child_by_handle' from public to protected,
since it doesn't appear to be needed for the public interface.
And added method 'nullify_db', which is called by
'update_parent_after_handle_change'. It was added because of
the new precondition for 'db=' that the argument not be nil.

RedisSTodoManager: Changed 'store_target' to, instead of
calling 'prepare_for_db_write' (which is now private) and
then calling '@data_manager.store_target', check if
'! tgt.connected_to_database'. If tgt is not connected,
it calls tgt.db= to set the 'db' attribute to @data_manager.
Then it calls tgt.force_update to let the STodoTarget update
itself and maintain its own integrity.
The same change was also made to 'save_new_targets'.

In runtests, added the new test 'remove-desc-101'.