New 'clone' test and corresponding changes/fixes
clone test: clone-101; added it to runtests.

Clone implementation changes:
First, there was a change in the spec, reflected by the test: The
the new/cloned STodoTarget is now expected to have the same parent
as the original STodoTarget. That is, it will have the same
'parent_handle' and the clone will be added as one of the parent's
children. STodoTarget.initialize_copy was updated to implement this
spec. Note that the new clone needs to be added as a child to the
parent outside of 'initialize_copy' since it's not a good idea to
have the STodoTarget object doing a database query for the parent,
nor to cause an update by calling add_child. Instead 'make_clone'
(STodoTargetEditor) retrieves the parent and calls:
        parent.add_child(clone)
Note also that the 'update' call at the end of initialize_copy was
removed - The update needs to be done (or caused, or forced) outside
of the STodoTarget after changing the handle to ensure the handles
are not the same. This is done in STodoTargetEditor.make_clone.

Due to the new spec that the parent of the clone is inherited from
the original, a 'make_orphan_clone' method was added that implements
the original, "orphaned" behavior. This method has not been tested.

Also, in RedisBasedDataManager.has_key?, changed the implementation
to return a boolean rather than nil or an object:
   def has_key?(handle)
-    result = key_table[handle]
+    result = ! key_table[handle].nil?
   end