Began: server and CLI client using redis
Began work on using more classes and modules borrowed from the
"tradable-analysis-toolset": Subscriber, Publisher, PublisherSubscriber,
Service, etc. The server subscribes to client requests on the
"SERVER_REQUEST_CHANNEL" and creates a Worker instance (which will at
some point run asynchronously) that carries out the client's request
and then sends a response. The client publishes its request and will
(when finished) subscribe to the server's response.

This version is working, but largely stub code.