Continue with server/client work; client REPL
Worker constructor (used by STodoServer) now takes the global
Configuration object as its only argument. It then creates the
TemplateTargetBuilder (which needs the Configuration object) used
for processing/creating new STodoTarget objects and stores it via
the 'manager' (which Worker also creates), which it passes to the
AddCommand/WorkCommand constructor. The benefit with respect to
this change is that since the Worker object (eventually, there
will be a "pool" of them) will not constantly be created and
destroyed as the server processes commands, the
TemplateTargetBuilder, which is rather heavy, will not constantly
be reconstructed. (The original design, which I think was not
committed, had the AddCommand/WorkCommand creating the
TemplateTargetBuilder, which is wasteful since the AddCommand (and
other commands that might use TemplateTargetBuilder) is created
for each client request. To facilitate the "reuse" of a
TemplateTargetBuilder object, TemplateTargetBuilder and its parent
TargetBuilder gained a new method:
  def spec_collector=(sc)
The 'spec_collector' attribute, which is used to obtain the
new attributes for the new STodoTarget item, needs to be
reset before each 'add' operation. This was originally done
in the TemplateTargetBuilder constructor, which meant a
TemplateTargetBuilder needed to be created to service each
client request. The performance appears to be much better with
this change.

RedisPubSubBroker and Subscription: Changed the name of
'subscribe_once' by making 'once' an argument in
RedisPubSubBroker.  The command: redis.unsubscribe channel
is now executed only if 'once' is true. (It defaults to 'false'.)

With these changes, the server now processes 'add' commands
(add task/memo/appointment/project) sent by the client to
add a new STodoTarget object.

Created a client REPL script, in addition to the CLI client
(stodo_cli_client) script.