command architecture change, implementations
Slight change to the structure of server command creation and
execution. The table of command-creator lambdas in CommandFacilities
was changed to a table of already created commands, so that a
new command does not have to be created for each execution.
Correspondingly, the 'execute' method of the command classes
was changed to take an argument: the 'request' object received
from the client.

Implemented ChangeCommand and ChangeHandleCommand. Fixed
DeleteCommand.

TargetBuilder: Added methods 'set_edit_mode' and 'set_create_mode'.
set_edit_mode is used in ChangeCommand.execute.