Handing off of response work to "WorkServer"s
First attempt at a change in the server structure to coordinate
and delegate the work needed for client responses to "WorkServer"s.
A WorkServer is ordered to fulfill a request via publish/subscribe.
It tells the object in charge (WorkCoordinator) that it is finished
via redis.set.

These changes include temporary code to write output to log files
for debugging purposes, since I could not figure out how to
attach the ruby debugger to a process running in the background.
I left the code in for now, since more debugging may be needed.