Add client session.
The client (STodoCliREPL) now requests a client session, which
will, among other attributes, contains the user-id, app-name,
and session-id. On the server side, a new command is "plugged
in" to handle these requests: SessionRequestCommand, which
creates the session, using the userid and appname passed by
the user in the client session request and assigning a
'new_session_id'. It then sends the new 'session' object back
to the client via the 'send_session' mechanism described
below under "WorkCommand...".

WorkCommand and descendants: Found a justification to make the
'execute' method in WorkCommand concrete, not abstract. 'execute'
now takes "the_caller" (i.e., the object that calls 'execute')
as its argument, rather than the (client) request object.
The caller has a 'request' query, which is used to set the new
'request' attribute (of WorkCommand) to the request object.
The reason for passing "the caller" to 'execute' is for a callback
facility. Specifically, the new SessionRequestCommand calls
the_caller.send_session to send the new ClientSession object to
the client. ("The caller" is either going to be Worker [for the
single-process testing version] or WorkServer.)

Also: A slight modification to RedisBasedDataManager; and a
cleaner way to set up the rubylib paths so that the client
and server processes can find the needed *.rb files.