First complete version of sessions
WorkServer functions as a delegate worker that responds to
client requests (delegated by the main/coordinating server).
Worker is the single-process (i.e., only one instance, running
in the main server process) version of this. The two classes
have much in common, so refactored to use a module that does the
common work (ClientRequestHandler).

database (RedisBasedDataManager): Since setting of the important
attribute 'db_key' is dependent on 'app_name' and 'user' (as is
'key_for', which prepend the app_name and user to the passed-in key),
added a 'set_appname_and_user' method to set those attributes as
a unit. It is also called by ClientRequestHandler to set the app_name
and user to the values received from the client request/session.
Changed the constructor to use this method.

Some corresponding changes were also needed for the client
STodoCliREPL.